refactor: extract content router (PR-6, Phase 1)
